Moved all censor-related functions in a separate file.
