+++ util.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"crypto/sha256"
 	_ "embed"
-	"encoding/base64"
 	"encoding/hex"
 	"encoding/json"
 	"math"
@@ -14,7 +13,6 @@ import (
 	"text/tabwriter"
 	"time"
 
-	goaway "github.com/TwiN/go-away"
 	"github.com/acarl005/stripansi"
 	markdown "github.com/quackduck/go-term-markdown"
 )
@@ -284,29 +282,6 @@ func devbotRespond(room *Room, messages []string, chance int) {
 func shasum(s string) string {
 	h := sha256.Sum256([]byte(s))
 	return hex.EncodeToString(h[:])
-}
-
-var detector = goaway.NewProfanityDetector().WithSanitizeSpaces(false)
-
-func rmBadWords(text string) string {
-	if !Config.Censor {
-		return text
-	}
-	return detector.Censor(text)
-}
-
-func init() {
-	okayIshWords := []string{"ZnVjaw==", "Y3JhcA==", "c2hpdA==", "YXJzZQ==", "YXNz", "YnV0dA==", "cGlzcw=="} // base 64 encoded okay-ish swears
-	for i := 0; i < len(goaway.DefaultProfanities); i++ {
-		for _, okayIshWord := range okayIshWords {
-			okayIshWordb, _ := base64.StdEncoding.DecodeString(okayIshWord)
-			if goaway.DefaultProfanities[i] == string(okayIshWordb) {
-				goaway.DefaultProfanities = append(goaway.DefaultProfanities[:i], goaway.DefaultProfanities[i+1:]...)
-				i-- // so we don't skip the next word
-				break
-			}
-		}
-	}
 }
 
 func holidaysCheck(u *User) {
